requirements.txt
groq>=0.11.0
langchain>=0.3.0
langgraph>=0.2.0
langchain-groq>=0.2.0
langchain-community>=0.3.0
langchain-core>=0.3.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
pydantic>=2.0.0
rich>=13.7.0
numpy>=1.24.3
typing-extensions>=4.5.0
